Xmgrace accelerator
The notebook helps to accelerate the post-MD visualization using pandas, seaborn and pyplot. The user needs to place the ipynb in the folder in which the subfolders' name are the name of the small molecules. In each subfolder, the xvg files should have the same name, such as rmsd_p100.xvg; SASA.xvg,...